programa 5,15,18,19
