+++ sumaElementos(libre19).cpp
@@ -1,7 +1,7 @@
  /*Realice una función que tome como parámetros un vector de números enteros
 y devuelva la suma de sus elementos.*/
 
-#include<iostream>
+/*#include<iostream>
 #include<conio.h>
 using namespace std;
 
@@ -36,4 +36,29 @@ int calcularSuma(int vec[],int tam){
 		suma += vec[i];
 	}
 	return suma;
+}*/
+#include<iostream>
+using namespace std;
+
+void pedirDatos(int vec[], int &tam);
+int calcularSuma(int vec[], int tam);
+
+void pedirDatos(int vec[], int &tam) {
+    cout << "Digite el numero de elementos del arreglo: ";
+    cin >> tam;
+    
+    for (int i = 0; i < tam; i++) {
+        cout << i + 1 << ". Digite un numero: ";
+        cin >> vec[i];
+    }
 }
+
+int calcularSuma(int vec[], int tam) {
+    int suma = 0;
+    
+    for (int i = 0; i < tam; i++) {
+        suma += vec[i];
+    }
+    return suma;
+}
+
